Migrate to local data instead of using TrailAPI

+++ hiking/backend-hiking/models/album_hikes_model.py
@@ -1,6 +1,5 @@
-from sqlalchemy import Column, String, Float, ForeignKey
-from sqlalchemy.orm import relationship, declarative_base
-from models.user_albums_model import User_Albums
+from sqlalchemy import Column, String, Integer
+from sqlalchemy.orm import declarative_base
 
 Base = declarative_base()
 metadata = Base.metadata
@@ -8,11 +7,8 @@ metadata = Base.metadata
 class Album_Hikes(Base):
         __tablename__ = "album_hikes"
         hike_id = Column(String(255), nullable=False, primary_key=True)
-        hike_api_id = Column(Float)
+        hike_api_id = Column(Integer)
         album_id = Column(String(255), nullable=False, primary_key=True)
-
-        # user_albums = relationship('User_Albums', back_populates='album_hikes')
-        # album_hikes = relationship('Hikes', back_populates='album_hikes')
 
         def __init__(self, hike_id, hike_api_id, album_id):
                 self.hike_id = hike_id
